model_creation/requirements.txt
entrypoints==0.3
enum34==1.1.10
flatbuffers==1.12
gast==0.4.0
h5py==3.1.0
httplib2==0.20.2
idna==3.4
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==4.8.1
imutils==0.5.4
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
jeepney==0.7.1
joblib==1.2.0
jsonschema==4.0.1
keras==2.6.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==23.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
moderngl==5.6.4
moderngl-window==2.4.0
mpmath==1.2.1
multipledispatch==0.6.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nlp2==1.8.38
notebook==6.4.4
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.2
Pillow==8.2.0
pkginfo==1.7.1
proto-plus==1.19.2
protobuf==3.18.0
ptyprocess==0.7.0
pyparsing==3.0.9
pyrr==0.10.3
requests==2.28.2
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.2.1
scipy==1.7.1
screeninfo==0.6.7
SecretStorage==3.3.1
sympy==1.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-advanced-segmentation-models @ git+https://github.com/JanMarcelKezmann/TensorFlow-Advanced-Segmentation-Models.git@c1cfb14840d8f3673c85f7989ece63d5c832605b
tensorflow-estimator==2.5.0
tensorflow-examples===6ae97eaf3dbd607ed3eccf18f7dc05d7a3b677e3-
tensorflow-model-optimization==0.7.3
tensorflow-objec
terminado==0.12.1
tesseract==0.1.3
testpath==0.5.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2021.8.30
tqdm==4.62.3
uritemplate==4.1.1
urllib3==1.26.14
validators==0.18.2
